django_microservice/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pika==1.3.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
zipp==3.14.0
